update testUtils to match other packages
