+++ packages/apollos-ui-connected/src/index.js
@@ -1,3 +1,3 @@
 export { LiveProvider, LiveConsumer } from './live';
 export LikedContentList, { GET_LIKED_CONTENT } from './LikedContentList';
-export { fetchMoreResolver, renderWithApolloData } from './utils';
+export { fetchMoreResolver } from './utils';
