+++ packages/apollos-ui-connected/src/utils/index.js
@@ -1,2 +1,1 @@
 export fetchMoreResolver from './fetchMoreResolver';
-export renderWithApolloData from './renderWithApolloData';
